requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.29
supabase>=2.0.2
requests>=2.28
beautifulsoup4>=4.12
APScheduler>=3.10.0
python-dotenv>=1.0
pandas>=2.1
matplotlib>=3.8
streamlit-extras
pydantic
email-validator